finish matrixPatchSubdiv

+++ src/thirdpartyplugins/filter_subdivfitting/filter_subdivfitting.h
@@ -78,7 +78,7 @@ private:
 	Eigen::VectorXd weightsPatch(const CFaceO* ft, float v, float w);
 	Eigen::VectorXd               weightsIrregularPatch(int V, float v, float w);
 	Eigen::RowVectorXd            weightsRegularPatch(float u, float v, float w);
-	Eigen::MatrixXi               matrixPickUP(int N, int k);
+	Eigen::MatrixXd               matrixPickUP(int N, int k);
 	Eigen::MatrixXd               matrixPatchSubdiv(int N);
 	float                         eps = 1.f / 64.f;
 };
